--- a/src/Microsoft.AspNetCore.Server.Kestrel/Internal/Http/FrameHeaders.Generated.cs
+++ b/src/Microsoft.AspNetCore.Server.Kestrel/Internal/Http/FrameHeaders.Generated.cs
@@ -5853,7 +5853,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Internal.Http
 
         protected override void SetValueFast(string key, StringValues value)
         {
-            ValidateHeaderCharacters(value);
+            ValidateHeaderCharacters(ref value);
             switch (key.Length)
             {
                 case 13:
@@ -6158,7 +6158,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Internal.Http
 
         protected override bool AddValueFast(string key, StringValues value)
         {
-            ValidateHeaderCharacters(value);
+            ValidateHeaderCharacters(ref value);
             switch (key.Length)
             {
                 case 13: