--- a/tools/Microsoft.AspNetCore.Server.Kestrel.GeneratedCode/KnownHeaders.cs
+++ b/tools/Microsoft.AspNetCore.Server.Kestrel.GeneratedCode/KnownHeaders.cs
@@ -372,7 +372,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Internal.Http
 
         protected override void SetValueFast(string key, StringValues value)
         {{{(loop.ClassName == "FrameResponseHeaders" ? @"
-            ValidateHeaderCharacters(value);" : "")}
+            ValidateHeaderCharacters(ref value);" : "")}
             switch (key.Length)
             {{{Each(loop.HeadersByLength, byLength => $@"
                 case {byLength.Key}: