@benaadams 2017-01-23
@cesarbs found the benchmark start issue; if you include a src project in global.json; you also need to include it in the Performance global.json otherwise it compiles, rerestores, compiles and first test errors as compile hasn't written when it starts. Is a bit weird