@cesarbs 2017-01-23
@benaadams Are you referring to https://github.com/aspnet/KestrelHttpServer/pull/1314#issuecomment-274579228? Problem is we don't have `global.json` anymore.