@benaadams 2017-01-23
One in https://github.com/aspnet/KestrelHttpServer/blob/dev/test/Microsoft.AspNetCore.Server.Kestrel.Performance/global.json